complete register and login

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Http\Request;
+use App\Http\Middleware\myAuth;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,23 +19,40 @@ Route::get('/user', function (Request $request) {
 })->middleware('auth:api');
 
 
-Route::get('news',  [ 'as' => 'news.index', 'uses' => 'newsController@index'/*, 'middleware' => 'auth:api']*/]);
-Route::get('news/{id}',  [ 'as' => 'news.show', 'uses' => 'newsController@show' /*, 'middleware' => 'auth:api']*/]);
-Route::get('news/create',  [ 'as' => 'news.create', 'uses' => 'newsController@create']);
-Route::get('news/{id}/edit',  [ 'as' => 'news.create', 'uses' => 'newsController@edit']);
-Route::post('news',  [ 'as' => 'news.create', 'uses' => 'newsController@store']);
-Route::put('news/{id}',  [ 'as' => 'news.create', 'uses' => 'newsController@update']);
-//Route::patch('news/{id}',  [ 'as' => 'news.create', 'uses' => 'newsController@update']);
-Route::delete('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@destroy']);
-Route::put('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@delete']);
-Route::put('news/{id}/restore',  [ 'as' => 'news.index', 'uses' => 'newsController@restore']);
+Route::get('news',  [ 'as' => 'news.index', 'uses' => 'newsController@index']);
+Route::get('news/{id}',  [ 'as' => 'news.show', 'uses' => 'newsController@show']);
 Route::get('search', [ 'as' => 'news.search', 'uses' => 'newsController@search']);
 
+Route::get('news/create',  [ 'as' => 'news.create', 'uses' => 'newsController@create', 'middleware' =>myAuth::class]);
+Route::get('news/{id}/edit',  [ 'as' => 'news.create', 'uses' => 'newsController@edit', 'middleware' =>myAuth::class]);
+Route::post('news',  [ 'as' => 'news.create', 'uses' => 'newsController@store', 'middleware' =>myAuth::class]);
+Route::put('news/{id}',  [ 'as' => 'news.create', 'uses' => 'newsController@update', 'middleware' =>myAuth::class]);
+//Route::patch('news/{id}',  [ 'as' => 'news.create', 'uses' => 'newsController@update']);
+Route::delete('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@destroy', 'middleware' =>myAuth::class]);
+Route::put('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@delete', 'middleware' =>myAuth::class]);
+Route::put('news/{id}/restore',  [ 'as' => 'news.index', 'uses' => 'newsController@restore', 'middleware' =>myAuth::class]);
 
-Route::post('login',  [ 'as' => '', 'uses' => 'Auth\LoginController@index']);
-Route::post('register',  [ 'as' => '', 'uses' => 'Auth\RegisterController@create'])/*->with( 'request', $request )*/;
 
-// Route::group(['prefix'=>'api','middleware'=>'auth:api'], function(){
-//    Route::get('news',  [ 'as' => 'news.index', 'uses' => 'newsController@index']);
+
+Route::post('login',  [ 'as' => '', 'uses' => 'Auth\LoginController@login']);
+Route::get('logout',  [ 'as' => '', 'uses' => 'Auth\LoginController@logout', 'middleware' =>myAuth::class]);
+Route::post('register',  [ 'as' => '', 'uses' => 'Auth\RegisterController@create_api']);
+
+
+
+// Route::group(['prefix'=>'api','middleware'=>myAuth::class], function(){
+// 	Route::get('news',  [ 'as' => 'news.index', 'uses' => 'newsController@index']);
+//     Route::post('news',  [ 'as' => 'news.create', 'uses' => 'newsController@store']);
+// 	Route::put('news/{id}',  [ 'as' => 'news.create', 'uses' => 'newsController@update']);
+// 	Route::delete('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@destroy']);
+//     Route::put('news/{id}',  [ 'as' => 'news.index', 'uses' => 'newsController@delete']);
+//     Route::put('news/{id}/restore',  [ 'as' => 'news.index', 'uses' => 'newsController@restore']);
 // });
 
+Route::get('user',  [ 'as' => 'user.index', 'uses' => 'userController@index']);
+Route::get('user/{id}',  [ 'as' => 'user.show', 'uses' => 'userController@show']);
+Route::get('user/{id}/edit',  [ 'as' => 'user.create', 'uses' => 'userController@edit']);
+Route::put('user/{id}',  [ 'as' => 'user.create', 'uses' => 'userController@update']);
+Route::delete('user/{id}',  [ 'as' => 'user.index', 'uses' => 'userController@destroy']);
+Route::put('user/{id}',  [ 'as' => 'user.index', 'uses' => 'userController@delete']);
+Route::put('user/{id}/restore',  [ 'as' => 'user.index', 'uses' => 'userController@restore']);
